+++ src/tests/slave_tests.cpp
@@ -3320,7 +3320,9 @@ TEST_F(SlaveTest, PingTimeoutNoPings)
   AWAIT_READY(slaveRegisteredMessage);
   ASSERT_TRUE(slaveRegisteredMessage->has_connection());
   MasterSlaveConnection connection = slaveRegisteredMessage->connection();
-  EXPECT_EQ(totalTimeout, Seconds(connection.total_ping_timeout_seconds()));
+  EXPECT_EQ(
+      totalTimeout,
+      Seconds(static_cast<int64_t>(connection.total_ping_timeout_seconds())));
 
   // Ensure the slave processes the registration message and schedules
   // the ping timeout, before we advance the clock.
